// src/app/shared/models/puesto-control.model.ts
export class PuestoControl {
    idPuestoControl: Number;
    nombrePuestoControl: string;
    controlObligatorio: string;
    departamento: string;
    provincia: string;
    distrito: string;
    coordenadasNorte: string;
    coordenadasEste: string;
    zonaUTM: Number;
    idAtf: Number;
    estado: string
}